添加会议人员

+++ smart-security/src/main/java/com/st/smartsecurity/pojo/po/MeetingManager.java
@@ -7,25 +7,25 @@ public class MeetingManager {
     @Id
     @Column(name = "`id`")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @Column(name = "`meeting_id`")
     private Long meetingId;
 
     @Column(name = "`manager_id`")
-    private Long managerId;
+    private Integer managerId;
 
     /**
      * @return id
      */
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
     /**
      * @param id
      */
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
 
@@ -46,14 +46,14 @@ public class MeetingManager {
     /**
      * @return manager_id
      */
-    public Long getManagerId() {
+    public Integer getManagerId() {
         return managerId;
     }
 
     /**
      * @param managerId
      */
-    public void setManagerId(Long managerId) {
+    public void setManagerId(Integer managerId) {
         this.managerId = managerId;
     }
 }
